# [CANN] Support Q4_0 for Ascend NPU