@fan-chao 2024-08-13
Does this commit support for Ascend310B1 chip?