@fan-chao 2024-08-13
> @fan-chao, it can`t work for Ascend310B1 card currently. Only support training card, such as Atlas 300T A2.

Ok，thanks.